+++ CHANGELOG.md
@@ -3,12 +3,9 @@
 ## 4.0.0 - 2023-11-XX
 
 ### Added
-<<<<<<< HEAD
 - Optionally serve a did:web DID document
 - Optionally serve a did-configuration document with DomainLinkageCredentials.
-=======
 - Microsoft Entra Verified ID exchange type
->>>>>>> 78f57aa (Protect from crashing if auth undefined)
 - New configFrom property to enable relying party inheritance.
 
 ### Fixed
@@ -17,6 +14,7 @@
 - Made ID Regex in endpoints less restrictive.
 - Config change to use consistent property casing. Requires config update!
 - Moved results into proper step.
+- Removed unused 'custom' exchange type.
 
 ## 3.0.2 - 2023-11-03